pyproject.toml
[project]
name = "hotel-booking-service"
version = "0.1.0"
description = "Simple REST API service for managing hotel rooms and bookings."
authors = [
    { name = "Pankrat", email = "panknotdeadd@gmail.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "django (>=6.0.2,<7.0.0)",
    "typing-extensions (>=4.15.0,<5.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "filelock (==3.24.3)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 100
target-version = "py313"
src = ["src"]
extend-exclude = [".venv", "venv", "build", "dist", "__pycache__", "migrations", "manage.py"]


[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "W", "I", "UP", "B", "SIM", "DJ", "PT"]
ignore = []


[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]
combine-as-imports = true
force-sort-within-sections = true


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
docstring-code-format = true


[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
warn_unused_configs = true
check_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
exclude = ".*migrations.*"


[dependency-groups]
dev = [
    "ruff (>=0.15.6,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "bandit (>=1.9.4,<2.0.0)",
    "radon (>=6.0.1,<7.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "ipython (>=9.11.0,<10.0.0)",
    "django-extensions (>=4.1,<5.0)"
]

test = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)"
]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
pythonpath = ["src"]